+++ features/Templates/library/image-banner/manifest.php
@@ -15,6 +15,7 @@
 	'areas' => [
 		'content' => [
 			'label' => 'Banner content',
+			'labelKey' => '/_admin/templates/area/banner-content',
 			'help' => 'The ordered content displayed over the image.',
 			'source' => 'single',
 			'allowed' => [ 'title', 'subtitle', 'description', 'button', 'template' ],